+++ graal-js/src/com.oracle.truffle.js/src/com/oracle/truffle/js/runtime/JSRuntime.java
@@ -56,6 +56,7 @@ import com.oracle.truffle.api.object.ObjectType;
 import com.oracle.truffle.api.profiles.BranchProfile;
 import com.oracle.truffle.api.profiles.ConditionProfile;
 import com.oracle.truffle.js.lang.JavaScriptLanguage;
+import com.oracle.truffle.js.runtime.array.TypedArrayFactory;
 import com.oracle.truffle.js.runtime.builtins.JSAbstractArray;
 import com.oracle.truffle.js.runtime.builtins.JSAdapter;
 import com.oracle.truffle.js.runtime.builtins.JSArray;
@@ -2803,5 +2804,9 @@ public final class JSRuntime {
     @SuppressWarnings("unchecked")
     public static <E extends Throwable> RuntimeException rethrow(Throwable ex) throws E {
         throw (E) ex;
+    }
+
+    public static boolean isTypedArrayBigIntFactory(TypedArrayFactory factory) {
+        return factory == TypedArrayFactory.BigInt64Array || factory == TypedArrayFactory.BigUint64Array;
     }
 }